only install ffmpeg if its not already present

+++ bin/custom-init.sh
@@ -2,7 +2,9 @@
 
 # see: https://hub.docker.com/r/lmscommunity/logitechmediaserver
 
-apt-get update -qq
-apt-get install --no-install-recommends -qy ffmpeg
+if [ ! -f /usr/bin/ffmpeg ] ; then
+    apt-get update -qq
+    apt-get install --no-install-recommends -qy ffmpeg
+fi
 
 exit 0
